DAO implementation added

+++ CoffeeShopIDEAProject/src/main/java/dao/IngridientDAO.java
@@ -2,7 +2,7 @@ package dao;
 
 import model.Ingridient;
 
-public interface IngridientDAO extends GenericDAO<Ingridient, Long>{
+public interface IngridientDAO extends GenericDAO<Ingridient, Integer>{
 
     boolean ingridientExist(String username);
